+++ packages/contracts-bedrock/contracts/L2/GasPriceOracle.sol
@@ -23,48 +23,6 @@ import { L1Block } from "../L2/L1Block.sol";
  */
 contract GasPriceOracle is Semver {
     /**
-     * @custom:legacy
-     * @custom:spacer _owner
-     * @notice Spacer for backwards compatibility.
-     */
-    address private spacer_0_0_20;
-
-    /**
-     * @custom:legacy
-     * @custom:spacer gasPrice
-     * @notice Spacer for backwards compatibility.
-     */
-    uint256 private spacer_1_0_32;
-
-    /**
-     * @custom:legacy
-     * @custom:spacer l1BaseFee
-     * @notice Spacer for backwards compatibility.
-     */
-    uint256 private spacer_2_0_32;
-
-    /**
-     * @custom:legacy
-     * @custom:spacer overhead
-     * @notice Spacer for backwards compatibility.
-     */
-    uint256 private spacer_3_0_32;
-
-    /**
-     * @custom:legacy
-     * @custom:spacer scalar
-     * @notice Spacer for backwards compatibility.
-     */
-    uint256 private spacer_4_0_32;
-
-    /**
-     * @custom:legacy
-     * @custom:spacer decimals
-     * @notice Spacer for backwards compatibility.
-     */
-    uint256 private spacer_5_0_32;
-
-    /**
      * @notice Number of decimals used in the scalar.
      */
     uint256 public constant DECIMALS = 6;